When sshing Only send specified public keys
$ man ssh_config
IdentitiesOnly
Specifies that ssh(1) should only use the authentication identity files
configured in the ssh_config files, even if ssh-agent(1) or a PKCS11Provider
offers more identities.  The argument to this keyword must be ``yes'' or
``no''.  This option is intended for situations where ssh-agent offers many
different identities.  The default is ``no''.